+++ packages/compiler/src/server/serverlib.ts
@@ -134,7 +134,7 @@ import {
   ServerSourceFile,
   ServerWorkspaceFolder,
 } from "./types.js";
-import { UpdateManger } from "./update-manager.js";
+import { UpdateManager } from "./update-manager.js";
 
 export function createServer(
   host: ServerHost,
@@ -161,7 +161,7 @@ export function createServer(
     (exports) => exports.$linter !== undefined,
   );
 
-  const updateManager = new UpdateManger(log);
+  const updateManager = new UpdateManager(log);
 
   const compileService = createCompileService({
     fileService,
@@ -1098,11 +1098,14 @@ export function createServer(
     };
     const resolved = await resolveModule(host, importPath, {
       baseDir: getDirectoryPath(currentFile.file.path),
+      directoryIndexFiles: ["main.tsp", "index.mjs", "index.js"],
       resolveMain(pkg) {
         // this lets us follow node resolve semantics more-or-less exactly
         // but using tspMain instead of main.
         return resolveTspMain(pkg) ?? pkg.main;
       },
+      conditions: ["typespec"],
+      fallbackOnMissingCondition: true,
     });
     return {
       uri: fileService.getURL(resolved.type === "file" ? resolved.path : resolved.mainFile),